Removed unused items
Realized that none of the metrics have any of the "switch_type" strings in them (i.e. "set", "toggle", "enable", etc.), so no switches are created; only "cabinsetpoint" was ever created, and that was incorrect.

I think that switches need to be created more explicitly based on the relevant OVMS commands.